refactor: move items shared by both ipfix and pcap backends in capture.rs
